Fix some issues with new index.php
Fix problems revealed on a different CDash testing environment.
