+++ models/buildgroup.php
@@ -15,7 +15,7 @@
      PURPOSE.  See the above copyright notices for more information.
 
 =========================================================================*/
-include_once('build.php');
+include_once('models/build.php');
 
 class BuildGroup
 {
